+++ BootloaderCorePkg/Include/Library/StageLib.h
@@ -7,6 +7,8 @@
 
 #ifndef _STAGE_LIB_H_
 
+#include <BootloaderCoreGlobal.h>
+
 /**
   Load IDT table for current processor.
